+++ Changelog.md
@@ -37,6 +37,7 @@ and this project adheres to
 - MEMBER uses EQL as the predicate to match number values.
   Otherwise working with character value lists would become
   a rather hairy issue.
+- AND, LAST, OR, and SYMBOL issue an error for dotted pairs.
 
 
 ## [v0.0.2] - 2024-08-17